componetize upload recording and teacherquestiontimeline
